requirements.txt
flask>=3.0
feedparser>=6.0
trafilatura>=1.8
deepl>=1.17
python-docx>=1.1
requests>=2.31
